+++ tegral-featureful/src/main/kotlin/guru/zoroark/tegral/featureful/ConfigurableFeature.kt
@@ -20,7 +20,7 @@ import guru.zoroark.tegral.config.core.ConfigurationSection
  * Represents a feature that provides configuration sections. These sections will be automatically parsed from Tegral
  * configuration files within the `[tegral]` section.
  */
-interface ConfigurableFeature : Feature {
+interface ConfigurableFeature<T> : Feature<T> {
     /**
      * Returns the sections that this feature provides for the `[tegral]` section.
      */